feat: :alembic: Adding springboot-webflux and all elements related to the benchmarking (grafana, prometheus, telegraf, docker-compose, etc...)

+++ rust/src/schemas/stats_response.rs
@@ -6,6 +6,6 @@ use crate::models::{priority::Priority, ticket_status::TicketStatus};
 
 #[derive(Serialize)]
 pub struct StatsResponse {
-    pub by_status: HashMap<TicketStatus, i32>,
-    pub by_priority: HashMap<Priority, i32>,
+    pub by_status: HashMap<TicketStatus, i64>,
+    pub by_priority: HashMap<Priority, i64>,
 }
